--- src/main/java/duke/View.java
@@ -1,101 +1,0 @@
-package duke;
-
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.Region;
-import javafx.scene.layout.VBox;
-import javafx.stage.Stage;
-
-public class View {
-    private ScrollPane scrollPane;
-    private VBox dialogContainer;
-    private TextField userInput;
-    private Button sendButton;
-    
-    /**
-     * Setting up required components for the elements displayed by view
-     * @param stage the primary stage for this application, onto which the application scene can be set, which
-     *  is pased in by Duke.
-     * @return Region is the base class for all JavaFX Node-based UI Controls, and all layout containers. Return
-     *  the region serving as the root node for the application.
-     */
-    public Region initialize() {
-        /** scrollPane is a container that displays its contents as a scrollable list. */
-        scrollPane = new ScrollPane();
-        /** Contents are displayed as a vertical list. */
-        dialogContainer = new VBox();
-        scrollPane.setContent(dialogContainer);
-
-        userInput = new TextField();
-        sendButton = new Button("Send");
-    
-        /** mainLayout is the root node which nests all the other elements in the view. */
-        AnchorPane mainLayout = new AnchorPane();
-        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
-
-        return mainLayout;
-    }
-
-    // Formats the window to standard specifications.
-    public void format(Stage stage, Region mainLayout) {
-        stage.setTitle("Duke");
-        stage.setResizable(false);
-        stage.setMinHeight(600.0);
-        stage.setMinWidth(400.0);
-        
-        mainLayout.setPrefSize(400.0, 600.0);
-        
-        scrollPane.setPrefSize(385, 535);
-        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
-        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
-        
-        scrollPane.setVvalue(1.0);
-        scrollPane.setFitToWidth(true);
-        
-        // You will need to import `javafx.scene.layout.Region` for this. 
-        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
-        
-        userInput.setPrefWidth(325.0);
-        
-        sendButton.setPrefWidth(55.0);
-        
-        AnchorPane.setTopAnchor(scrollPane, 1.0);
-        
-        AnchorPane.setBottomAnchor(sendButton, 1.0);
-        AnchorPane.setRightAnchor(sendButton, 1.0);
-        
-        AnchorPane.setLeftAnchor(userInput , 1.0);
-        AnchorPane.setBottomAnchor(userInput, 1.0);
-    }
-
-    public void setEventListeners() {
-        //Step 3. Add functionality to handle user input.
-        sendButton.setOnMouseClicked((event) -> {
-            dialogContainer.getChildren().add(getDialogLabel(userInput.getText()));
-            userInput.clear();
-        });
-        
-        userInput.setOnAction((event) -> {
-            dialogContainer.getChildren().add(getDialogLabel(userInput.getText()));
-            userInput.clear();
-        });
-    }
-        
-
-    /**
-     * Iteration 1:
-     * Creates a label with the specified text and adds it to the dialog container.
-     * @param text String containing text to add
-     * @return a label with the specified text that has word wrap enabled.
-     */
-    private Label getDialogLabel(String text) {
-        // You will need to import `javafx.scene.control.Label`.
-        Label textToAdd = new Label(text);
-        textToAdd.setWrapText(true);
-
-        return textToAdd;
-    }
-}
